refactor predicate to its own package
