+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -14,9 +14,9 @@ import java.util.function.Predicate;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.TagsContainsTagPredicate;
+import seedu.address.model.person.predicate.NameContainsKeywordsPredicate;
+import seedu.address.model.person.predicate.TagsContainsTagPredicate;
 import seedu.address.model.tag.AbstractTag;
 
 /**
