+++ src/main/java/seedu/address/model/person/predicate/TagsContainsTagPredicate.java
@@ -1,9 +1,10 @@
-package seedu.address.model.person;
+package seedu.address.model.person.predicate;
 
 import java.util.Set;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.model.person.Person;
 import seedu.address.model.tag.AbstractTag;
 
 /**
